// src/types.ts
export type ScanType = 'JOB_OFFER' | 'RENTAL_SCAM';
export type InputType = 'TEXT' | 'URL' | 'PDF' | 'IMAGE';